index calc outsourced cc transmit

+++ assignment4/meltdown-template.c
@@ -97,7 +97,7 @@ static void cc_setup() {
 
 // cc_transmit(uint8_t value) transmits a value through the channel
 static void cc_transmit(uint8_t value) {
-    int index = cc_buf_offset+value*512+value;
+    u64 index = cc_buf_offset+value*512+value;
     my_mfence();
     maccess(&cc_buffer[index]);
     my_mfence();
